requirements.txt
python-telegram-bot==20.7
python-docx==1.1.0
docx2pdf==0.1.8